Working on removing hardhat-deploy
