+++ abi/diamond-1-hardhat/contracts/upgradeInitializers/DiamondInit.sol/DiamondInit.json
@@ -1,0 +1,9 @@
+[
+  {
+    "inputs": [],
+    "name": "init",
+    "outputs": [],
+    "stateMutability": "nonpayable",
+    "type": "function"
+  }
+]
