+++ contracts/registry/facets/PCOLicenseClaimerFacet.sol
@@ -6,9 +6,8 @@ import "../libraries/LibPCOLicenseParams.sol";
 import "../../pco-license/facets/CFABasePCOFacet.sol";
 import "../interfaces/IPCOLicenseParamsStore.sol";
 import {IERC721} from "@solidstate/contracts/interfaces/IERC721.sol";
-import "hardhat-deploy/solc_0.8/diamond/libraries/LibDiamond.sol";
-import "../../beacon-diamond/BeaconDiamond.sol";
-import {IDiamondLoupe} from "hardhat-deploy/solc_0.8/diamond/interfaces/IDiamondLoupe.sol";
+import {BeaconDiamond} from "../../beacon-diamond/BeaconDiamond.sol";
+import {IDiamondLoupe} from "diamond-1-hardhat/contracts/interfaces/IDiamondLoupe.sol";
 import {IConstantFlowAgreementV1} from "@superfluid-finance/ethereum-contracts/contracts/interfaces/agreements/IConstantFlowAgreementV1.sol";
 import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import "../../beneficiary/interfaces/ICFABeneficiary.sol";
