--- deployments/optimism-goerli/.chainId
@@ -1,1 +1,0 @@
-420
