+++ package.json
@@ -26,10 +26,11 @@
     "@truffle/hdwallet-provider": "^1.1.1",
     "bn.js": "^5.1.3",
     "date-fns": "^2.28.0",
-    "hardhat-diamond-abi": "^3.0.1",
-    "remixd": "^0.2.4-alpha.0"
+    "remixd": "^0.2.4-alpha.0",
+    "diamond-1-hardhat": "https://github.com/mudgen/diamond-1-hardhat#639662e7e9b81b6595f7e8b95f44a1f5ba86dd41"
   },
   "devDependencies": {
+    "hardhat-diamond-abi": "^3.0.1",
     "@defi-wonderland/smock": "^2.0.7",
     "@nomiclabs/hardhat-ethers": "npm:hardhat-deploy-ethers",
     "@nomiclabs/hardhat-waffle": "^2.0.3",
