+++ scripts/deploy.ts
@@ -1,0 +1,84 @@
+import SuperfluidSDK from "@superfluid-finance/js-sdk";
+const deployFramework = require("@superfluid-finance/ethereum-contracts/scripts/deploy-framework");
+const deploySuperToken = require("@superfluid-finance/ethereum-contracts/scripts/deploy-super-token");
+import { Framework, SuperToken } from "@superfluid-finance/sdk-core";
+const hre = require("hardhat");
+import { Contract } from "ethers";
+import { ethers } from "hardhat";
+
+async function deploySuperfluid() {
+  const errorHandler = (err: any) => {
+    if (err) throw err;
+  };
+
+  const { diamondAdmin } = await hre.getNamedAccounts();
+
+  await deployFramework(errorHandler, {
+    web3: hre.web3,
+    from: diamondAdmin,
+  });
+
+  await deploySuperToken(errorHandler, [":", "ETH"], {
+    web3: hre.web3,
+    from: diamondAdmin,
+  });
+
+  const jsSf = new SuperfluidSDK.Framework({
+    web3: hre.web3,
+    version: "test",
+    tokens: ["ETH"],
+  });
+  await jsSf.initialize();
+
+  const sf = await Framework.create({
+    chainId: hre.network.config.chainId,
+    provider: hre.web3,
+    resolverAddress: (jsSf.resolver as Contract).address,
+    protocolReleaseVersion: "test",
+  });
+
+  const ethx = await sf.loadSuperToken(jsSf.tokens.ETHx.address);
+
+  return { sf, ethx };
+}
+
+async function deployDiamond() {
+  // name: string,
+  // facetNames: string[],
+  // from: string,
+  // owner: string
+  // Deploy DiamondInit
+  // DiamondInit provides a function that is called when the diamond is upgraded or deployed to initialize state variables
+  // Read about how the diamondCut function works in the EIP2535 Diamonds standard
+  const DiamondInit = await ethers.getContractFactory("DiamondInit");
+  const diamondInit = await DiamondInit.deploy();
+  await diamondInit.deployed();
+  console.log("DiamondInit deployed:", diamondInit.address);
+}
+
+async function deploy() {
+  let sf: Framework;
+  let ethx: SuperToken;
+  if (hre.network.config.chainId == 31337) {
+    const res = await deploySuperfluid();
+    sf = res.sf;
+    ethx = res.ethx;
+  } else {
+    sf = await Framework.create({
+      chainId: hre.network.config.chainId!,
+      provider: hre.ethers.provider,
+    });
+    ethx = await sf.loadSuperToken("ETHx");
+  }
+
+  await deployDiamond();
+}
+
+if (require.main === module) {
+  deploy()
+    .then(() => process.exit(0))
+    .catch((error) => {
+      console.error(error);
+      process.exit(1);
+    });
+}
